feat: centralize env int helper
