+++ alpha_factory_v1/backend/agents/cyber_threat_agent.py
@@ -101,20 +101,13 @@ except ModuleNotFoundError:  # pragma: no cover
 from backend.agent_base import AgentBase  # pylint: disable=import‑error
 from backend.agents import AgentMetadata, register_agent
 from backend.orchestrator import _publish  # reuse orchestrator event bus
+from alpha_factory_v1.utils.env import _env_int
 
 logger = logging.getLogger(__name__)
 
 # ---------------------------------------------------------------------------
 # Configuration structure
 # ---------------------------------------------------------------------------
-
-
-def _env_int(name: str, default: int) -> int:
-    try:
-        return int(os.getenv(name, default))
-    except ValueError:
-        return default
-
 
 def _env_float(name: str, default: float) -> float:
     try:
