+++ alpha_factory_v1/backend/agents/energy_agent.py
@@ -103,17 +103,13 @@ except ModuleNotFoundError:  # pragma: no cover
 from backend.agents.base import AgentBase  # pylint: disable=import-error
 from backend.agents import AgentMetadata, register_agent
 from backend.orchestrator import _publish  # reuse event-bus helper
+from alpha_factory_v1.utils.env import _env_int
 
 logger = logging.getLogger(__name__)
 
 
 # ---------------------------------------------------------------------------
 # Configuration ----------------------------------------------------------------
-def _env_int(name: str, default: int) -> int:
-    try:
-        return int(os.getenv(name, default))
-    except ValueError:
-        return default
 
 
 @dataclass
