+++ alpha_factory_v1/backend/agents/retail_demand_agent.py
@@ -91,20 +91,13 @@ except ModuleNotFoundError:  # pragma: no cover
 from backend.agent_base import AgentBase  # pylint: disable=import-error
 from backend.agents import AgentMetadata, register_agent
 from backend.orchestrator import _publish  # structured‑event helper
+from alpha_factory_v1.utils.env import _env_int
 
 logger = logging.getLogger(__name__)
 
 # ---------------------------------------------------------------------------
 # Configuration dataclass  ---------------------------------------------------
 # ---------------------------------------------------------------------------
-
-
-def _env_int(name: str, default: int) -> int:
-    try:
-        return int(os.getenv(name, default))
-    except ValueError:
-        return default
-
 
 def _env_bool(name: str) -> bool:
     return os.getenv(name, "").lower() in {"1", "true", "yes"}
