+++ alpha_factory_v1/utils/env.py
@@ -6,6 +6,10 @@ from __future__ import annotations
 import os
 from pathlib import Path
 from typing import Dict
+import logging
+
+
+log = logging.getLogger(__name__)
 
 
 def _load_env_file(path: str | os.PathLike[str]) -> Dict[str, str]:
@@ -28,3 +32,16 @@ def _load_env_file(path: str | os.PathLike[str]) -> Dict[str, str]:
         k, v = line.split("=", 1)
         data[k.strip()] = v.strip().strip('"')
     return data
+
+
+def _env_int(name: str, default: int) -> int:
+    """Return ``int`` environment value or ``default`` if conversion fails."""
+
+    val = os.getenv(name)
+    if val is None:
+        return default
+    try:
+        return int(val)
+    except (TypeError, ValueError):
+        log.warning("Invalid %s=%r, using default %s", name, val, default)
+        return default
